+++ .github/actions/build-rpms/build-rpms.sh
@@ -28,9 +28,12 @@ if [[ "$GITHUB_REF" == refs/pull/*/merge ]]; then
 	pr_number="$(<<<"$GITHUB_REF" grep -o '[0-9]\+')"
 	files_changed="$(curl "${GITHUB_API_URL}/repos/${GITHUB_REPOSITORY}/pulls/${pr_number}/files" | jq -r .[].filename)"
 else
-	files_changed="$(git diff-tree --no-commit-id --name-only -r "$GITHUB_SHA")"
+	files_changed="$(git diff --name-only HEAD~4 HEAD --)" # Files changed by the last 4 commits
 fi
-if <<<"$files_changed" grep '^GO_VERSION$'; then
+if <<<"$files_changed" grep '^GO_VERSION$' ||
+	{ [[ "$ATC_COMPONENT" == traffic_router ]] &&
+		<<<"$files_changed" grep '^infrastructure/docker/build/Dockerfile-traffic_router$'; };
+then
 	pkg_command+=(-b)
 fi