+++ cache-config/testing/ort-tests/tcdata/parameters.go
@@ -1,3 +1,5 @@
+package tcdata
+
 /*
 
    Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,13 +15,11 @@
    limitations under the License.
 */
 
-package tcdata
-
 import (
 	"sync"
 	"testing"
 
-	tc "github.com/apache/trafficcontrol/lib/go-tc"
+	"github.com/apache/trafficcontrol/lib/go-tc"
 )
 
 func (r *TCData) CreateTestParameters(t *testing.T) {
@@ -61,26 +61,26 @@ func DeleteTestParameter(t *testing.T, pl tc.Parameter) {
 	// Retrieve the Parameter by name so we can get the id for the Update
 	resp, _, err := TOSession.GetParameterByNameAndConfigFile(pl.Name, pl.ConfigFile)
 	if err != nil {
-		t.Errorf("cannot GET Parameter by name: %v - %v", pl.Name, err)
+		t.Errorf("cannot GET Parameter by name %s: %v", pl.Name, err)
 	}
 
 	if len(resp) == 0 {
 		// TODO This fails for the ProfileParameters test; determine a way to check this, even for ProfileParameters
-		// t.Errorf("DeleteTestParameter got no params for %+v %+v", pl.Name, pl.ConfigFile)
+		// t.Errorf("DeleteTestParameter got no params for %s %s", pl.Name, pl.ConfigFile)
 	} else if len(resp) > 1 {
 		// TODO figure out why this happens, and be more precise about deleting things where created.
-		// t.Errorf("DeleteTestParameter params for %+v %+v expected 1, actual %+v", pl.Name, pl.ConfigFile, len(resp))
+		// t.Errorf("DeleteTestParameter params for %s %s expected 1, actual %d", pl.Name, pl.ConfigFile, len(resp))
 	}
 	for _, respParameter := range resp {
 		delResp, _, err := TOSession.DeleteParameterByID(respParameter.ID)
 		if err != nil {
-			t.Errorf("cannot DELETE Parameter by name: %v - %v", err, delResp)
+			t.Errorf("cannot DELETE Parameter by ID: %v - %v", err, delResp)
 		}
 
 		// Retrieve the Parameter to see if it got deleted
 		pls, _, err := TOSession.GetParameterByID(pl.ID)
 		if err != nil {
-			t.Errorf("error deleting Parameter name: %s", err.Error())
+			t.Errorf("error deleting Parameter name: %v", err)
 		}
 		if len(pls) > 0 {
 			t.Errorf("expected Parameter Name: %s and ConfigFile: %s to be deleted", pl.Name, pl.ConfigFile)