+++ cache-config/testing/ort-tests/tcdata/todb.go
@@ -1,3 +1,5 @@
+package tcdata
+
 /*
 
    Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,12 +15,9 @@
    limitations under the License.
 */
 
-package tcdata
-
 import (
 	"database/sql"
 	"fmt"
-	"os"
 
 	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/auth"
@@ -36,70 +35,64 @@ func (r *TCData) OpenConnection() (*sql.DB, error) {
 		sslStr = "disable"
 	}
 
-	fmt.Printf("User: %s, Password: len %d, Hostname: %s, DB: %s\n", r.Config.TrafficOpsDB.User, len(r.Config.TrafficOpsDB.Password), r.Config.TrafficOpsDB.Hostname, r.Config.TrafficOpsDB.Name)
+	log.Debugf("User: %s, Password: len %d, Hostname: %s, DB: %s\n", r.Config.TrafficOpsDB.User, len(r.Config.TrafficOpsDB.Password), r.Config.TrafficOpsDB.Hostname, r.Config.TrafficOpsDB.Name)
 	db, err = sql.Open("postgres", fmt.Sprintf("postgres://%s:%s@%s/%s?sslmode=%s", r.Config.TrafficOpsDB.User, r.Config.TrafficOpsDB.Password, r.Config.TrafficOpsDB.Hostname, r.Config.TrafficOpsDB.Name, sslStr))
 
 	if err != nil {
 		log.Errorf("opening database: %v\n", err)
-		return nil, fmt.Errorf("transaction failed: %s", err)
+		return nil, fmt.Errorf("transaction failed: %w", err)
 	}
 	return db, err
 }
 
-// SetupTestData ...
-// TODO error does not need returned as this function can never return a non-nil error
+// SetupTestData sets up initial data needed by the test suite to perform any
+// tests.
 func (r *TCData) SetupTestData(*sql.DB) error {
-	var err error
-
-	err = SetupRoles(db)
+	err := SetupRoles(db)
 	if err != nil {
-		fmt.Printf("\nError setting up roles %s - %s, %v\n", r.Config.TrafficOps.URL, r.Config.TrafficOps.Users.Admin, err)
-		os.Exit(1)
+		return fmt.Errorf("setting up Roles: %w", err)
 	}
 
 	err = SetupCapabilities(db)
 	if err != nil {
-		fmt.Printf("\nError setting up capabilities %s - %s, %v\n", r.Config.TrafficOps.URL, r.Config.TrafficOps.Users.Admin, err)
-		os.Exit(1)
+		return fmt.Errorf("setting up capabilities: %w", err)
 	}
 
 	err = SetupRoleCapabilities(db)
 	if err != nil {
-		fmt.Printf("\nError setting up roleCapabilities %s - %s, %v\n", r.Config.TrafficOps.URL, r.Config.TrafficOps.Users.Admin, err)
-		os.Exit(1)
+		return fmt.Errorf("setting up roleCapabilities: %w", err)
 	}
 
 	err = SetupAPICapabilities(db)
 	if err != nil {
-		fmt.Printf("\nError setting up APICapabilities %s - %s, %v\n", r.Config.TrafficOps.URL, r.Config.TrafficOps.Users.Admin, err)
-		os.Exit(1)
+		return fmt.Errorf("setting up APICapabilities: %w", err)
 	}
 
 	err = SetupTenants(db)
 	if err != nil {
-		fmt.Printf("\nError setting up tenant %s - %s, %v\n", r.Config.TrafficOps.URL, r.Config.TrafficOps.Users.Admin, err)
-		os.Exit(1)
+		return fmt.Errorf("setting up tenants: %w", err)
 	}
 
 	err = r.SetupTmusers(db)
 	if err != nil {
-		fmt.Printf("\nError setting up tm_user %s - %s, %v\n", r.Config.TrafficOps.URL, r.Config.TrafficOps.Users.Admin, err)
-		os.Exit(1)
+		return fmt.Errorf("setting up tm_user: %w", err)
 	}
 
 	err = SetupTypes(db)
 	if err != nil {
-		fmt.Printf("\nError setting up types %s - %s, %v\n", r.Config.TrafficOps.URL, r.Config.TrafficOps.Users.Admin, err)
-		os.Exit(1)
+		return fmt.Errorf("setting up types: %w", err)
 	}
 
 	err = SetupToExtensions(db)
 	if err != nil {
-		fmt.Printf("\nError setting up to_extensions %s - %s, %v\n", r.Config.TrafficOps.URL, r.Config.TrafficOps.Users.Admin, err)
-		os.Exit(1)
+		return fmt.Errorf("setting up to_extensions: %w", err)
 	}
 
-	return err
+	return nil
+}
+
+func execError(sql string, err error) error {
+	return fmt.Errorf("encountered error '%w' during execution of query: %s", err, sql)
 }
 
 // SetupRoles ...
@@ -116,7 +109,7 @@ INSERT INTO role (name, description, priv_level) VALUES ('federation','Role for 
 `
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -132,7 +125,7 @@ INSERT INTO capability (name, description) VALUES ('cache-groups-read', 'Read CG
 `
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -146,7 +139,7 @@ INSERT INTO api_capability (http_method, route, capability) VALUES ('GET', '/cac
 
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -158,7 +151,7 @@ INSERT INTO role_capability (role_id, cap_name) VALUES (4,'all-read') ON CONFLIC
 `
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -169,7 +162,7 @@ func (r *TCData) SetupTmusers(db *sql.DB) error {
 	var err error
 	encryptedPassword, err := auth.DerivePassword(r.Config.TrafficOps.UserPassword)
 	if err != nil {
-		return fmt.Errorf("password encryption failed %v", err)
+		return fmt.Errorf("password encryption failed %w", err)
 	}
 
 	// Creates users in different tenants
@@ -184,7 +177,7 @@ INSERT INTO tm_user (username, local_passwd, role, tenant_id) VALUES ('` + r.Con
 `
 	err = execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -200,7 +193,7 @@ INSERT INTO tenant (name, active, parent_id, last_updated) VALUES ('badtenant', 
 `
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -213,7 +206,7 @@ INSERT INTO deliveryservice_tmuser (deliveryservice, tm_user_id, last_updated) V
 `
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -228,7 +221,7 @@ INSERT INTO job (id, ttl_hr, asset_url, start_time, entered_time, job_user, last
 `
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -243,7 +236,7 @@ INSERT INTO type (name, description, use_in_table) VALUES ('CHECK_EXTENSION_OPEN
 `
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -257,7 +250,7 @@ INSERT INTO to_extension (name, version, info_url, isactive, script_file, server
 	`
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return nil
 }
@@ -309,7 +302,7 @@ func (r *TCData) Teardown(db *sql.DB) error {
 `
 	err := execSQL(db, sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v", err)
+		return execError(sqlStmt, err)
 	}
 	return err
 }
@@ -320,17 +313,17 @@ func execSQL(db *sql.DB, sqlStmt string) error {
 
 	tx, err := db.Begin()
 	if err != nil {
-		return fmt.Errorf("transaction begin failed %v %v ", err, tx)
+		return fmt.Errorf("transaction begin failed: %w", err)
 	}
 
 	res, err := tx.Exec(sqlStmt)
 	if err != nil {
-		return fmt.Errorf("exec failed %v %v", err, res)
+		return execError(sqlStmt, err)
 	}
 
 	err = tx.Commit()
 	if err != nil {
-		return fmt.Errorf("commit failed %v %v", err, res)
+		return fmt.Errorf("transaction commit failed %w %v", err, res)
 	}
 	return nil
 }