+++ traffic_ops/traffic_ops_golang/crstats/routing.go
@@ -207,7 +207,6 @@ func getCRSStats(respond chan<- RouterResp, wg *sync.WaitGroup, routerFQDN, cdn 
 
 // getCDNRouterFQDNs returns an FQDN, including port, of an online router for each CDN, for each router. If a CDN has no online routers, that CDN will not have an entry in the map. The port returned is the API port.
 func getCDNRouterFQDNs(tx *sql.Tx, requiredCDN *string) (map[tc.CDNName][]string, error) {
-	var args interface{}
 	query := `
 SELECT s.host_name, s.domain_name, max(pa.value) as port, c.name as cdn
 FROM server as s
@@ -222,12 +221,17 @@ AND st.name = '` + RouterOnlineStatus + `'
 `
 	if requiredCDN != nil {
 		query += `AND c.name = $1`
-		args = *requiredCDN
 	}
 	query += `
 GROUP BY s.host_name, s.domain_name, c.name
 `
-	rows, err := tx.Query(query, args)
+	var rows *sql.Rows
+	var err error
+	if requiredCDN != nil {
+		rows, err = tx.Query(query, *requiredCDN)
+	} else {
+		rows, err = tx.Query(query)
+	}
 	if err != nil {
 		return nil, errors.New("querying routers: " + err.Error())
 	}